asr/models/conv_rec_framework/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py38h7f8727e_2
alembic=1.7.6=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
asteroid-filterbanks=0.4.0=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
autopage=0.5.0=pypi_0
backports-cached-property=1.0.1=pypi_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h497a2fe_1001
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38hd667e15_1
chardet=4.0.0=py38h578d9bd_2
charset-normalizer=2.0.11=pyhd8ed1ab_0
click=8.0.3=py38h578d9bd_1
cliff=3.10.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=2.3.3=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
colorlog=6.6.0=pypi_0
cryptography=35.0.0=py38ha5dfef3_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pypi_0
docopt=0.6.2=pypi_0
einops=0.3.2=pypi_0
freetype=2.10.4=h0708190_1
fsspec=2022.1.0=pyhd8ed1ab_0
future=0.18.2=py38h578d9bd_4
google-auth=2.6.0=pyh6c4a22f_1
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
greenlet=1.1.2=pypi_0
grpcio=1.42.0=py38hce63b2e_0
icu=67.1=he1b5a44_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.0=py38h578d9bd_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pypi_0
julius=0.2.6=pypi_0
kiwisolver=1.3.1=py38h2531618_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=h21135ba_2
libprotobuf=3.15.8=h780b84a_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libuv=1.40.0=h7b6447c_0
llvmlite=0.38.0=pypi_0
mako=1.1.6=pypi_0
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=pypi_0
matplotlib-base=3.2.2=py38h5d868c9_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpmath=1.2.1=pypi_0
multidict=5.2.0=py38h7f8727e_2
ncurses=6.3=h7f8727e_2
networkx=2.6.3=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.2.0=pyhd8ed1ab_0
omegaconf=2.1.1=pypi_0
openssl=1.1.1m=h7f8727e_0
optuna=2.10.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.2.3=py38h51da96c_0
pbr=5.8.1=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.4=py38h06a4308_0
prettytable=3.1.0=pypi_0
primepy=1.3=pypi_0
protobuf=3.15.8=py38h709712a_0
pyannote-audio=0.0.1=pypi_0
pyannote-database=4.1.3=pypi_0
pyannote-metrics=3.2=pypi_0
pyannote-pipeline=2.2=pypi_0
pyannote.core=4.3=py38h578d9bd_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pydeprecate=0.3.1=pypi_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyperclip=1.8.2=pypi_0
pysocks=1.7.1=py38h578d9bd_4
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.2=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.5.10=pyhd8ed1ab_0
pytorch-metric-learning=1.1.1=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=5.4.1=py38h497a2fe_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py38hc147768_0
semver=2.13.0=pypi_0
setuptools=59.5.0=pypi_0
shellingham=1.4.0=pypi_0
simplejson=3.17.6=py38h7f8727e_0
singledispatchmethod=1.0=pypi_0
six=1.16.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
spectralcluster=0.2.4=pypi_0
sqlalchemy=1.4.31=pypi_0
sqlite=3.37.2=hc218d9a_0
stevedore=3.5.0=pypi_0
sympy=1.9=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py38h2b97feb_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch-audiomentations=0.10.0=pypi_0
torch-pitch-shift=1.2.0=pypi_0
torchaudio=0.10.2=py38_cu113
torchmetrics=0.7.2=pyhd8ed1ab_0
torchvision=0.11.2=pypi_0
tornado=6.1=py38h497a2fe_1
tqdm=4.62.3=pyhd8ed1ab_0
typer=0.4.0=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pyhd8ed1ab_1
wcwidth=0.2.5=pypi_0
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py38h27cfd23_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h7f8727e_4
